pyproject.toml
[project]
name = "crawlerutils"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "tiramission", email = "jaign@outlook.com" }
]
dependencies = [
    "loguru>=0.7.2",
    "httpx>=0.27.0",
    "pyyaml>=6.0.1",
    "aiohttp>=3.9.5",
    "tenacity>=8.5.0",
]
readme = "README.md"
requires-python = ">= 3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
cache_dir = ".cache/pytest"
asyncio_default_fixture_loop_scope = "module"

[dependency-groups]
dev = [
      "pytest>=8.2.2",
    "pytest-asyncio>=0.23.8",
    "nox>=2024.10.9",
      
]

